progress

+++ tools/scripts/Generate-Spell.ps1
@@ -4,16 +4,21 @@
 param (
     [Parameter()]
     [string]
-    $Spell
+    $Spell,
+    $Element = "magic",
+    $Tier = "novice",
+    $Type = "attack",
+    $Cost = "10"
 )
 
 ###############################################
 # Static
 $script:srcPath = "C:\git\AncientSpellcraft\src"
 $script:modid = "ancientspellcraft"
+$script:spellRegistryFile = "C:\git\AncientSpellcraft\src\main\java\com\windanesz\ancientspellcraft\registry\ASSpells.java"
 ###############################################
 
-################ Item Specific ################ 
+################ Item Specific ################
 
 $script:spellName = $Spell
 
@@ -35,10 +40,10 @@ function Add-SpellJsonFile {
         "trades": true,
         "looting": true
     },
-    "tier": "novice",
-    "element": "magic",
-    "type": "attack",
-    "cost": 10,
+    "tier": "$Tier",
+    "element": "$Element",
+    "type": "$Type",
+    "cost": $Cost,
     "chargeup": 0,
     "cooldown": 20,
     "base_properties": {
@@ -47,8 +52,10 @@ function Add-SpellJsonFile {
 "@
 
     $filePath = "$($script:srcPath)\main\resources\assets\$($script:modid)\spells\$($script:spellName).json"
-    $json | Out-File $filePath -Encoding utf8 -NoNewline
-    Write-Host "Generated $filePath spell json"
+    if (!(Test-Path -Path $filePath)) {
+        $json | Out-File $filePath -Encoding utf8 -NoNewline
+        Write-Host "Generated $filePath spell json"
+    }
 }
 
 function Add-LangFileEntry {
@@ -65,8 +72,11 @@ function Create-Texture {
     $texturePath = $script:srcPath + "\main\resources\assets\$($script:modid)\textures\spells\$($script:spellName).png"
     if (!(Test-Path -Path $texturePath -PathType Leaf)) {
         $genericTexture = $script:srcPath + "\main\resources\assets\$($script:modid)\textures\spells\none.png"
-        Copy-Item $genericTexture -Destination $texturePath
-        Write-Host "Generated $texturePath texture"
+
+        if (!(Test-Path -Path $texturePath)) {
+            Copy-Item $genericTexture -Destination $texturePath
+            Write-Host "Generated $texturePath texture"
+        }
     }
 }
 
@@ -86,8 +96,23 @@ function Create-Texture {
 
 #endregion Functions
 
-Add-LangFileEntry -Entry "item.$($script:modid)\:$($script:spellName).name=TODO"
-Add-LangFileEntry -Entry "item.$($script:modid)\:$($script:spellName).desc=TODO"
+
+$displayName = (Get-Culture).TextInfo.ToTitleCase([String]::Join(" ", $spellName.Split("_")))
+Add-LangFileEntry -Entry "spell.$($script:modid)\:$($script:spellName).desc=TODO"
+Add-LangFileEntry -Entry "spell.$($script:modid)\:$($script:spellName)=$displayName"
 Create-Texture
 Add-SpellJsonFile
 
+# Add member
+$lineOfLastSpell = (Select-String -Path $script:spellRegistryFile -Pattern 'public\sstatic\sfinal\sSpell')[-1].LineNumber
+$textToAdd = "`n    public static final Spell $($script:spellName) = placeholder();"
+$fileContent = Get-Content $script:spellRegistryFile
+$fileContent[$lineOfLastSpell-1] += $textToAdd
+$fileContent | Set-Content $script:spellRegistryFile
+
+# Add registry placeholder
+$lineOfLastSpell = (Select-String -Path $script:spellRegistryFile -Pattern 'registry\.register\(new\s')[-1].LineNumber
+$textToAdd = "`n        registry.register(new <<ADD SPELL>>);"
+$fileContent = Get-Content $script:spellRegistryFile
+$fileContent[$lineOfLastSpell-1] += $textToAdd
+$fileContent | Set-Content $script:spellRegistryFile
